+++ ocpi211/tariffs.go
@@ -25,7 +25,7 @@ const (
 
 type PriceComponent struct {
 	Type     TariffDimensionType `json:"type"`
-	Price    float64             `json:"price"`
+	Price    float32             `json:"price"`
 	StepSize int                 `json:"step_size"`
 }
 
@@ -34,10 +34,10 @@ type TariffRestriction struct {
 	EndTime     *string         `json:"end_time,omitempty"`
 	StartDate   *string         `json:"start_date,omitempty"`
 	EndDate     *string         `json:"end_date,omitempty"`
-	MinKWh      *float64        `json:"min_kwh,omitempty"`
-	MaxKWh      *float64        `json:"max_kwh,omitempty"`
-	MinPower    *float64        `json:"min_power,omitempty"`
-	MaxPower    *float64        `json:"max_power,omitempty"`
+	MinKWh      *float32        `json:"min_kwh,omitempty"`
+	MaxKWh      *float32        `json:"max_kwh,omitempty"`
+	MinPower    *float32        `json:"min_power,omitempty"`
+	MaxPower    *float32        `json:"max_power,omitempty"`
 	MinDuration *int            `json:"min_duration,omitempty"`
 	MaxDuration *int            `json:"max_duration,omitempty"`
 	DayOfWeek   []DayOfWeekText `json:"day_of_week,omitempty"`
